rename to showmsg.[ch]
