+++ src/readqtccalls.c
@@ -29,7 +29,7 @@
 
 #include "qtcutil.h"
 #include "qtcvars.h"		// Includes globalvars.h
-#include "startmsg.h"
+#include "showmsg.h"
 #include "tlf_curses.h"
 
 int qtcdirection = 0;
