+++ src/showmsg.h
